Final cleanup of xterm.js, nothing but prototype left
